% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasePlot.R
\name{BasePlot}
\alias{BasePlot}
\title{An \code{\link{R6}} \code{BasePlot} class for \strong{\link{paramonte}} package usage}
\description{
This is the class for generating instances of basic plots with minimally one (X)-axis.
It serves as the superclass for a wide variety of other multi-axes ParaMonte plots.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BasePlot$new()}}
\item \href{#method-getLogLinSpace}{\code{BasePlot$getLogLinSpace()}}
\item \href{#method-clone}{\code{BasePlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePlot$new(
  plotType,
  dataFrame = NULL,
  methodName = "ParaMonte",
  reportEnabled = TRUE,
  resetPlot = NULL,
  isGrid = FALSE,
  Err
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLogLinSpace"></a>}}
\if{latex}{\out{\hypertarget{method-getLogLinSpace}{}}}
\subsection{Method \code{getLogLinSpace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePlot$getLogLinSpace(
  base,
  skip = 1,
  lowerLim = 1,
  upperLim = nrow(self$rows)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
