% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPlot1D.R
\name{DensityPlot1D}
\alias{DensityPlot1D}
\title{An \code{\link{R6}} \code{DensityPlot1D} class for \strong{\link{paramonte}} package usage}
\description{
This is the DensityPlot1D class for generating instances of histogram figures
in one dimension based on a wide range of plotting tools from the \code{ggplot}.\cr
Normally, the public users are not supposed to use this class directly, although they can
for the purposes other than plotting the ParaMonte simulation files.
}
\section{Super class}{
\code{\link[paramonte:BasePlot]{paramonte::BasePlot}} -> \code{DensityPlot1D}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DensityPlot1D$new()}}
\item \href{#method-reset}{\code{DensityPlot1D$reset()}}
\item \href{#method-make}{\code{DensityPlot1D$make()}}
\item \href{#method-clone}{\code{DensityPlot1D$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paramonte" data-topic="BasePlot" data-id="getLogLinSpace">}\href{../../paramonte/html/BasePlot.html#method-getLogLinSpace}{\code{paramonte::BasePlot$getLogLinSpace()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DensityPlot1D$new(
  plotType,
  dataFrame = NULL,
  methodName = "ParaMonte",
  reportEnabled = TRUE,
  resetPlot = NULL,
  isGrid = FALSE,
  Err
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DensityPlot1D$reset(resetType = "soft", isFirstTime = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make"></a>}}
\if{latex}{\out{\hypertarget{method-make}{}}}
\subsection{Method \code{make()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DensityPlot1D$make(reself = FALSE, xcolumns, show = TRUE, target = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DensityPlot1D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
