% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramonte_class.R
\name{paramonte_class}
\alias{paramonte_class}
\title{An \code{\link{R6}} \code{paramonte_class} for \strong{\link{paramonte}} package usage}
\description{
\code{paramonte_class$new()} initializes a \code{paramonte_class} object.\cr
It can then be used to instantiate \code{\link{ParaDRAM}} samplers object.\cr
}
\examples{

## ------------------------------------------------
## Method `paramonte_class$new`
## ------------------------------------------------

pm   = paramonte_class$new()

## ------------------------------------------------
## Method `paramonte_class$ParaDRAM`
## ------------------------------------------------

pm   = paramonte_class$new()
pmpd = pm$ParaDRAM()

## ------------------------------------------------
## Method `paramonte_class$print`
## ------------------------------------------------

pm = paramonte_class$new()
pm$print()
# or
pm = paramonte_class$new()
pm
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{paramonte_class$new()}}
\item \href{#method-ParaDRAM}{\code{paramonte_class$ParaDRAM()}}
\item \href{#method-print}{\code{paramonte_class$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instantiates a \code{paramonte_class} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a \code{\link{R6}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm   = paramonte_class$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParaDRAM"></a>}}
\if{latex}{\out{\hypertarget{method-ParaDRAM}{}}}
\subsection{Method \code{ParaDRAM()}}{
This method instantiate a \code{\link{R6}} \code{\link{ParaDRAM}} class object. \cr
It can then be used to call \code{\link{ParaDRAM}} routines. \cr
This is the \code{\link{ParaDRAM}} class constructor to generate instances of \strong{serial} and \strong{parallel}
\strong{Delayed-Rejection Adaptive Metropolis-Hastings Markov Chain Monte Carlo} sampler class
of the \code{\link{paramonte}} library. The \code{\link{ParaDRAM}} class is a child of the \code{\link{ParaMonteSampler}} class. \cr
The object of \code{\link{ParaDRAM}} class can be instantiated only via \code{\link{paramonte_class}} object method, \cr
(for example: \code{pm$ParaDRAM()}, where \code{pm = parmonte_classs$new()} ).\cr
All \code{\link{ParaDRAM}} class attributes are optional and all attributes can be set after a \code{\link{ParaDRAM}}
instance is returned by the class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$ParaDRAM()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a \code{\link{R6}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm   = paramonte_class$new()
pmpd = pm$ParaDRAM()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints \code{paramonte_class} object usage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{paramonte_class$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte_class$new()
pm$print()
# or
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{pm = paramonte_class$new()
pm
}
\if{html}{\out{</div>}}

}

}
}
