% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0printsum.R
\name{printsum}
\alias{printsum}
\title{My printsum function}
\usage{
printsum(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
returns sum of x and y
}
\description{
This function simply prints sum 2 numbers.
}
\examples{
sum(1, 2)
\dontrun{
printsum(4, 5)
}
}
