% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docParaMonteSampler.R
\name{readReport}
\alias{readReport}
\title{\code{readReport} method for \code{ParaDRAM} class objects for \link{paramonte} package}
\description{
\code{readReport} is a method for objects of class \code{\link{ParaDRAM}}.

click \href{../../paramonte/html/ParaDRAM.html#method-readChain}{here} for the documentation of \code{\link{ParaDRAM}} class to see
documentation of \code{readReport}.
}
\seealso{
\code{\link{readChain}} \code{\link{readSample}} \code{\link{readReport}} \code{\link{readRestart}} \code{\link{readProgress}} \code{\link{readMarkovChain}} \code{\link{ParaDRAM}}
}
